removed nested component -- it's a bad practice and put question context within the CardContent component
